Add StructuredNopLogger
